Create File to bootstrap an oracle container and run ks-with-rice-bundled
This script expects to be run on a computer setup with the docker daemon.

Then it will download the oracle image and start it.

Next it will run the impex process against the oracle container.

Next it will run the ks-with-rice-bundled application against the provisioned
oracle container.

It will run build-917 which was the final java version of kuali student.